moved named vector test utils into a sub directory under the  folder
