+++ test/acceptance_with_go_client/named_vectors_tests/test_suits/named_vectors_byov_with_vectorizer.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package named_vectors_tests
+package test_suits
 
 import (
 	"context"
