+++ test/acceptance_with_go_client/named_vectors_tests/test_suits/named_vectors_ref_props.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package named_vectors_tests
+package test_suits
 
 import (
 	"context"
